+++ src/main/java/com/fpt/gsu25se47/schoolpsychology/dto/response/SurveyRecordResponse.java
@@ -3,6 +3,7 @@ package com.fpt.gsu25se47.schoolpsychology.dto.response;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -12,7 +13,7 @@ public class SurveyRecordResponse {
 
     private Integer id;
     private String noteSuggest;
-    private Integer totalScore;
+    private BigDecimal totalScore;
     private String status;
     private String level;
     private LocalDate completedAt;
